Isolate segy ingestion leakage into base ingestion pipeline
