+++ src/mdio/ingestion/schema/__init__.py
@@ -1,20 +1,17 @@
-"""Immutable schema models that describe a dataset before any data is scanned.
+"""Immutable, format-agnostic schema models describing a dataset before any data is scanned.
 
-``SchemaResolver`` lives in :mod:`mdio.ingestion.schema.resolver` and is imported from there
-directly to avoid a circular import (the resolver depends on the SEG-Y index-strategy
-registry, which in turn depends on these models).
+These models (``ResolvedSchema``, ``DimensionSpec``) and the ``SchemaEffect`` reshape contract
+know nothing about any ingestion format. Concrete effects and the logic that selects them live
+with the format that needs them (for SEG-Y grid overrides, see
+:mod:`mdio.ingestion.segy.schema_effects`).
 """
 
-from mdio.ingestion.schema.models import CollapseToTraceEffect
 from mdio.ingestion.schema.models import DimensionSpec
-from mdio.ingestion.schema.models import InsertTraceDimEffect
 from mdio.ingestion.schema.models import ResolvedSchema
 from mdio.ingestion.schema.models import SchemaEffect
 
 __all__ = [
-    "CollapseToTraceEffect",
     "DimensionSpec",
-    "InsertTraceDimEffect",
     "ResolvedSchema",
     "SchemaEffect",
 ]
